FIXES: Create .env.ci file for DB Backup Workflow
WHY: GitHub Actions doesn't support passing env vars to docker compose,
so we need to create a .env.ci file for the DB Backup Workflow that
contains the DB credentials and pass that file to the docker compose
command.

Also added a check to validate the DB credentials are set.