Add script to upgrade composer dependencies
WHAT: called from docker-compose.yml internally in the php Dockerfile.
Will automatically run when the container starts and ensure composer
dependencies are up to date.

WHY: This is a simple way to ensure that composer dependencies are up to
date without having to manually run composer install or composer update
in the container.